+++ CHANGELOG.md
@@ -15,7 +15,8 @@
 ### Features
 - ODPresentation Writer : Show/Hide Value / Name / Series Name in Chart - @Progi1984 GH-272
 - ODPresentation Writer : Axis Bounds in Chart - @Progi1984 GH-269
-- PowerPoint2007 Reader : Support of Table - @Progi1984 GH-240
+- PowerPoint2007 Reader : Support of Shape Table - @Progi1984 GH-240
+- PowerPoint2007 Reader : Support of Slide Note - @Progi1984 GH-226
 - PowerPoint2007 Writer : Implement character spacing - @jvanoostrom GH-301
 - PowerPoint2007 Writer : Axis Bounds in Chart - @Progi1984 GH-269
 - PowerPoint2007 Writer : Implement Legend Key in Series for Chart  - @Progi1984 GH-319